+++ CHANGELOG.md
@@ -6,3 +6,5 @@
 * admin
   * show all procedures including Anträge which are voted on by parliament
   * show and link important documents including document number
+  * show history of procedures
+  * show findspot for Anträge